v0.0.3
Working on simulation:
The simulation program is actually complete,
I added initialization of the history (for both males, and separately
females, which is a little clunky but it seems like the best choice)
I gave history a current_turn variable so that when it’s passed into
the aviary, the birds know where to look
The response is completely centralized in the aviary, which takes
history and returns a turn.
History then records the turn in the proper place
This is repeated until all turns are taken.

This means that I need to write the respond function (and an “frespond”
and “mrespond” function, for male or female specific responses) which
lies in the aviary, which will simply call all the individual birds’
responses

By the way, normalizing the male investment is a little tricky, so
remember how that works (it requires reshaping matrices)