v0.0.6
Added dependencies!
Currently there are only two stats and two strategies (one male one
female).

-Stats are returned as an object, this allows for handy printing of
stats
-strategies return the whole investment matrix. Also remember that in
python everything is a pointer basically. I might be able to use that
to optimize a bunch of things, but it also means that I have to be
careful when changing stuff.

Based on what there is currently, everything should be ready to start
debug. Everything exists as nested functions, so I’ll need one more
program to handle everything, which will allow me to call the matingsim
in iPython (woo!)

a lot of debugging ahead…
After that, (I think we’ll call it v0.1 when it works) need to add
several more strategies and more stats.
As part of that, it would be great to show investment and reward like I
did in mathematica.