v0.1.0 debug
Starting debug. Major problems are fixed (import matingsim runs). Now I
can start working through individual programs in iPython.

Side Note: A separate window of iPython is a much better way to debug
than using the same window.

Prior to debugging, I also added a stat for the record (so now we have
history stats and record stats) as it is, I don’t actually remember if
it prints the stats after a simulation, I think I still need to connect
that, but that could be in a separate file.