v0.1.1 debug
Lots of debugging, my initial guess was way off. More like 20 typos and
50 bugs.

Fixed lots of bugs, and now it runs, so that’s a start. Currently
initial investment is 0, and subsequent investment is Nan. So we’ll
have to figure out what’s going on there. But a solid start.