v0.2.3 - Works like original
I played around with it for a while to recapitulate the original
results. It turns out that there’s a fairly critical component of the
male strategy: in order for monogamy to emerge, males must ignore 0’s
when computing their average reward. In other words, they only
calculate their benefit over the females they’re interacting with.
Otherwise, you end up with various, stable midpoints. I think this was
sort of an artifact of trying to remove 0 errors originally, but it
makes a fair amount of sense biologically.

Once I design some stats, this should work just like the original,
although it runs significantly faster.