V0.2.5 - It net-Works!
I added a network plotting function that will now put the progression
of the network over time, which is quite beautiful. As part of that, I
built an adjacency matrix function, which will segue very well into
using distance as a factor, which I think is the obvious next step. The
other next step, is to get the statistics back and have a good way to
plot that.