v0.3.3 - Adjacency platform working, strategy needs work
I’ve added basically a parallel simulation, which keeps a full
investment matrix and a reward matrix. I’ll need to rename several
things, but everything is working. As it stands, they invest based on
reward, and it all works, but they just bump around. The next steps are
to think about

-Strategy
-Parameters
-Mechanisms (make sure the investment - adj conversion makes sense)

But the program basically works. So that’s good.