v0.5.2 - Are you asking for a challenge?
- Basically, it all works. Monogamy occurs as a function of the decay
constant and the bump constant. So far alpha, iota, and beta don’t seem
to have a huge impact… need to look into that further.
- Competition as a mechanic is fully implemented*
*That said, the strategy for both males and females is to run away from
any competition (up to half of their resources at the moment). The cost
function is set such that the cost of being competed against is
actually fairly low, that could change, but it would mean I need to
think a little bit about what it means when your investment becomes
negative (death? roll over?), and whether it makes sense for
males/females to start out in random competition with each other (maybe
it doesn’t)