v0.6.1 - It works! And they're Monogamous!
I spent all day fiddling with parameters and mechanisms. Aside from a
bit of debugging, the biggest thing was figuring out a smooth
optimization strategy for males that allows them to get to a good
solution. Currently, males are good at this. Females are still just
picking a single male and going with it (which turns out to be a pretty
winning strategy for them). Next steps will be dealing with more
complex strategies (or just comparing to simpler strategies to see what
we can learn).