v0.6.2 - Playtime
Very small changes to track total reward for males and females as their
‘self’ reward. Also updated the female optimal strategy to work the
same as males.

Started some tests (see param_history.txt.) In short, monogamy is a
winning strategy for males and females in this universe (where theres a
significant cost to crowded females). Next steps: Continue tests, and
I’d like to add a strategy where males can only sing to one bird at a
time. This might solve the difficult of optimization (I doubt it) but
it will also lay the foundation for more realistic network behavior,
since our real birds have to invest as events not as this abstract
‘investment’ substance. It’s not exactly clear what that will look like
though…will need to think about it.